Refactor FigurJoeOptionListener
