--- src/test/java/edu/kis/powp/jobs2d/events/SelectTestFigureJoe1OptionListener.java
@@ -1,21 +1,0 @@
-package edu.kis.powp.jobs2d.events;
-
-import java.awt.event.ActionEvent;
-import java.awt.event.ActionListener;
-
-import edu.kis.powp.jobs2d.drivers.DriverManager;
-import edu.kis.powp.jobs2d.magicpresets.FiguresJoe;
-
-public class SelectTestFigureJoe1OptionListener implements ActionListener {
-
-	private DriverManager driverManager;
-
-	public SelectTestFigureJoe1OptionListener(DriverManager driverManager) {
-		this.driverManager = driverManager;
-	}
-
-	@Override
-	public void actionPerformed(ActionEvent e) {
-		FiguresJoe.figureScript1(driverManager.getCurrentDriver());
-	}
-}
