--- src/test/java/edu/kis/powp/jobs2d/events/SelectTestFigureJoe2OptionListener.java
@@ -1,20 +1,0 @@
-package edu.kis.powp.jobs2d.events;
-
-import edu.kis.powp.jobs2d.drivers.DriverManager;
-import edu.kis.powp.jobs2d.magicpresets.FiguresJoe;
-import java.awt.event.ActionEvent;
-import java.awt.event.ActionListener;
-
-public class SelectTestFigureJoe2OptionListener implements ActionListener {
-
-	private DriverManager driverManager;
-
-	public SelectTestFigureJoe2OptionListener(DriverManager driverManager) {
-		this.driverManager = driverManager;
-	}
-
-	@Override
-	public void actionPerformed(ActionEvent e) {
-		FiguresJoe.figureScript2(driverManager.getCurrentDriver());
-	}
-}
