+++ src/test/java/edu/kis/powp/jobs2d/events/TestCase.java
@@ -1,0 +1,5 @@
+package edu.kis.powp.jobs2d.events;
+
+public enum TestCase {
+    FIGURE_JOE1, FIGURE_JOE2;
+}
